requirements.txt
fastapi==0.111.0
pydantic==2.7.4
PyJWT==2.8.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
SQLAlchemy==2.0.30
starlette==0.37.2
uvicorn==0.30.1
